PlanVectorizer/requirements.txt
opencv-python
numpy
svgwrite
pillow